Li 2019 phenodata

List of phenodata per study. There are 3 studies with 4 branches:

s1

melanoma (nivolumab, phase1)

s2

RCC (nivolumab, phase1)

s3_nivo

RCC (nivolumab, phase3)

s3_ever

RCC (everolimus, phase3)

Each of the four list elements contains a data frame with samples as rows and pheno-information as columns.

Usage

li2019_pheno

Format

A list of 4 with a single tibble each.

s1

78 samples, 8 columns

s2

91 samples, 8 columns

s3_nivo

394 samples, 10 columns

s3_ever

349 samples, 10 columns

The data.frames s1 and s2 have the following columns:

Unique_Subject_Identifier

Identifier

Region

Sample region

Derived_Age_at_Consent

Age at consent

Sex

Sex. "M", "F".

Race

"WHITE" or "OTHER"

BRAF_V600E

BRAF_V600E mutation status

OS

Overall survival time

OS_Censor

Overall survival status

The data.frames s3_nivo and s3_ever have the following columns:

Unique_Subject_Identifier

Identifier

Treatment

"NIVOLUMAB" or "EVEROLIMUS"

Region

Sample region

Sex

"M" or "F"

Derived_Age_at_Consent

Age at consent

Race

"WHITE" or "OTHER"

CRF_MSKCC_Risk_Group

"FAVORABLE", "INTERMEDIATE" or "POOR"

Prior antiangiogenic regimens (≥2)

OS

Overall survival time

OS_Censor

Overall survival status

Source

Liu, D., Schilling, B., Liu, D. et al. Integrative molecular and clinical modeling of clinical outcomes to PD1 blockade in patients with metastatic melanoma. Nat Med 25, 1916–1927 (2019). https://doi.org/10.1038/s41591-019-0654-5